Fixing up errors and warnings >Removed error 4251 to avoid DLL style interfaces error >Marked _CRT_SECURE_NO_WARNINGS_ to stop throwing errors in IMGUI >Updated destruction framework
Minor Fixes
- AssetBase.cpp: sprintf_s now uses %zu to format size_t
- AssetBase.h: m_AssetSize is now initialised.
- Added files for Vector.h, InVector.h, Rotator.h
- Removed Frustum.h and all references.
- Removed duplicate BindMaterial definition since we now only use vertex buffer reference.
- Disabled D3D11 debug
- Fixed non-specified cast warnings

+++ Flow/Source/Flow/Window/WinWindow.cpp
@@ -217,13 +217,13 @@ namespace Flow
 			case WM_SYSKEYDOWN:
 			case WM_KEYDOWN:
 			{
-				KeyPressedEvent Event(wParam, 0);
+				KeyPressedEvent Event(static_cast<int>(wParam), 0);
 				e = &Event;
 				break;
 			}
 			case WM_KEYUP:
 			{
-				KeyReleasedEvent Event(wParam);
+				KeyReleasedEvent Event(static_cast<int>(wParam));
 				e = &Event;
 				break;
 			}
